+++ packages/components/src/border-box-control/border-box-control/component.tsx
@@ -118,7 +118,6 @@ const UnconnectedBorderBoxControl = (
 						__experimentalIsRenderedInSidebar={
 							__experimentalIsRenderedInSidebar
 						}
-						__shouldNotWarnDeprecated36pxSize
 						size={ size }
 					/>
 				) : (