+++ packages/components/src/border-control/border-control/hook.ts
@@ -12,7 +12,6 @@ import type { WordPressComponentProps } from '../../context';
 import { useContextSystem } from '../../context';
 import { useCx } from '../../utils/hooks/use-cx';
 import type { Border, BorderControlProps } from '../types';
-import { maybeWarnDeprecated36pxSize } from '../../utils/deprecated-36px-size';
 
 // If either width or color are defined, the border is considered valid
 // and a border style can be set as well.
@@ -37,20 +36,12 @@ export function useBorderControl(
 		value: border,
 		width,
 		__experimentalIsRenderedInSidebar = false,
-		__next40pxDefaultSize,
-		__shouldNotWarnDeprecated36pxSize,
+		__next40pxDefaultSize: _,
+		__shouldNotWarnDeprecated36pxSize: __,
 		...otherProps
 	} = useContextSystem( props, 'BorderControl' );
 
-	maybeWarnDeprecated36pxSize( {
-		componentName: 'BorderControl',
-		__next40pxDefaultSize,
-		size,
-		__shouldNotWarnDeprecated36pxSize,
-	} );
-
-	const computedSize =
-		size === 'default' && __next40pxDefaultSize ? '__unstable-large' : size;
+	const computedSize = size === 'default' ? '__unstable-large' : size;
 
 	const [ widthValue, originalWidthUnit ] = parseQuantityAndUnitFromRawValue(
 		border?.width
@@ -170,6 +161,5 @@ export function useBorderControl(
 		widthValue,
 		size: computedSize,
 		__experimentalIsRenderedInSidebar,
-		__next40pxDefaultSize,
 	};
 }