+++ packages/eslint-plugin/rules/components-no-missing-40px-size-prop.js
@@ -13,7 +13,6 @@ const { hasTruthyJsxAttribute } = require( '../utils' );
  */
 const COMPONENTS_REQUIRING_40PX = new Set( [
 	'BorderBoxControl',
-	'BorderControl',
 	'BoxControl',
 	'Button',
 	'ClipboardButton',