+++ packages/components/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Breaking Changes
 
 -   `TextControl`: The `__next40pxDefaultSize` prop is now true by default. The prop can be safely removed ([#79386](https://github.com/WordPress/gutenberg/pull/79386)).
+-   `BorderControl`: The `__next40pxDefaultSize` prop is now true by default. The prop can be safely removed.
 
 ### Enhancements